64b port
git-svn-id: svn+ssh://jhoberock-dt/home/jhoberock/svnbase/jared/src/gotham@1096 3bb10773-34f4-0310-b88c-ec9d511ba0d9

+++ shading/Fresnel.cpp
@@ -7,7 +7,7 @@
 #include "ScatteringDistributionFunction.h"
 
 void *Fresnel
-  ::operator new(unsigned int size)
+  ::operator new(size_t size)
 {
   return ScatteringDistributionFunction::mPool.malloc();
 } // end Fresnel::operator new()
